latihan_django/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
django-cors-headers==4.3.1
django-health-check==3.18.1
django-minio-backend==3.5.0
django-minio-storage==0.5.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.4
minio==7.2.3
openapi-codec==1.3.2
packaging==23.2
psutil==5.9.8
psycopg==3.1.17
psycopg2==2.9.9
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
django-filter==2.4.0
psycopg2-binary==2.9.9